Rename openai module
