+++ voc_builder/openai_svc.py
